move avatar url to its own module
